% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{extract_album_data}
\alias{extract_album_data}
\title{Extract Artist and Album Info from Artist Page}
\usage{
extract_album_data(url_artist)
}
\arguments{
\item{url_artist}{An artist URL, preferrably those extracted with \link{extract_artist_urls}.}
}
\description{
Scrape artist name and album data from an artist page as a data frame
}
\examples{
extract_album_data(url_artist = "http://www.progarchives.com/artist.asp?id=4398")
extract_album_data(url_artist = "http://www.progarchives.com/artist.asp?id=671")
}
