% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{extract_subgenres}
\alias{extract_subgenres}
\title{Extract All Subgenre URLs}
\usage{
extract_subgenres()
}
\description{
Scrape all subgenre names and URLs from \href{http://www.progarchives.com/}{ProgArchives} as a dataframe
}
\examples{
extract_subgenres()
}
