% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{scrape_progarchives}
\alias{scrape_progarchives}
\title{Scrape All Album Data from ProgArchives}
\usage{
scrape_progarchives(parallel = TRUE)
}
\arguments{
\item{parallel}{Logical value; should the function use parallelism when extracting album data? Defaults to TRUE.}
}
\description{
Extract subgenre, artist and album data from \href{http://www.progarchives.com/}{ProgArchives}
}
\examples{
\dontrun{
scrape_progarchives()
}
}
